requirements.txt
streamlit-ace
openai
primehub-python-sdk